Rust: Remove allocation from LED write
This commit removes the `to_vec()` calls from the Matrix thread. This is
achieved by instead using `into_iter().cloned()` which does not
allocate.

However, since we now are reading from the backend while writing to the
driver, we have to move all the logic into a single function. This
ensures that the borrow checker sees that `self.backend` and
`self.driver` are borrowed separately.

Signed-off-by: Sebastian Wicki <gandro@gmx.net>