Rust: Add own LED driver trait
This commit adds a new `WriteLeds` trait which replaces the
`SmartLedsWrite` trait used before. In contrast to the `SmartLedsWrite`
trait, which requires the caller to pass in an iterator, the new
`WriteLeds` trait accepts a slice of colors. This allows for zero-copy
driver implementations.

This commit also adds an adapter which converts a `SmartLedsWrite`
driver into a `WriteLeds` driver.

Signed-off-by: Sebastian Wicki <gandro@gmx.net>